requirements.txt
eve
flask-cors
